requirements.txt
django>=5.0
djangorestframework
python-dotenv
dj-database-url
psycopg2-binary
gunicorn
starkbank
djangorestframework-simplejwt
drf-spectacular
